requirements.txt
# requirements.txt
# Список зависимостей для запуска и проверки Flask-приложения (РГЗ, вариант 1)

# Flask — основной веб-фреймворк
flask

# Flasgger — автоматическая генерация Swagger UI и OpenAPI спецификации
flasgger

# PyYAML — преобразование JSON спецификации в формат OpenAPI (YAML)
pyyaml

# Bandit — инструмент статического анализа кода (проверка безопасности)
bandit
